+++ CollectionFile.cpp
@@ -164,7 +164,7 @@ CCollectionFile::~CCollectionFile()
 
 void CCollectionFile::WriteCollectionInfo(CFileDataIO *out_data)
 {
-	out_data->WriteUInt32(taglist.GetSize());
+	out_data->WriteUInt32((uint32)taglist.GetSize());
 
 	for (int i = 0; i < taglist.GetSize(); ++i) {
 		CTag tempTag(*taglist[i]);
