+++ MuleToolBarCtrl.h
@@ -18,7 +18,8 @@
 
 #define	MULE_TOOLBAR_BAND_NR	0
 
-enum EToolbarLabelType: int {
+enum EToolbarLabelType : uint8
+{
 	NoLabels	= 0,
 	LabelsBelow = 1,
 	LabelsRight = 2
